@@ -1,6 +1,6 @@
 module github.com/alex1989hu/kubelet-serving-cert-approver
 
-go 1.15
+go 1.16
 
 require (
 	github.com/cucumber/godog v0.11.0
